requirements.txt
pip==24.3.1
PyQtWebEngine==5.15.7
pywebview==5.3.2
QtPy==2.4.2
setuptools==75.2.0
tortoise-orm==0.21.7
wheel==0.44.0
